# api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Rest API created to manage ECU presets",
    "version": "1.0.0",
    "title": "Engine-Specs-API",
    "contact": {
      "email": "mcorpucci@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Engine Requests",
      "description": "Engine related specs"
    }
  ],
  "paths": {
    "/api/engines": {
      "get": {
        "tags": [
          "Engine Requests"
        ],
        "summary": "Returns the available Engine specs",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/EngineRequests"
            }
          }
        }
      }
    },
    "/api/engines/engine": {
      "post": {
        "security": {
          "bearerAuth": []
        },
        "tags": [
          "Engine Requests"
        ],
        "summary": "Create a new Engine request",
        "requestBody": {
          "description": "Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/engineRequestPostBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/engineResponseBody"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "$ref": "#/components/schemas/authResponseError"
          }
        }
      }
    },
    "/api/engines/engine/{engine_id}": {
      "get": {
        "summary": "Returns an Engine by its ID",
        "tags": [
          "Engine Requests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "engine_id",
            "required": true,
            "description": "Engine Request Id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/engineResponseBody"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an Engine by its ID",
        "security": {
          "bearerAuth": []
        },
        "tags": [
          "Engine Requests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "engine_id",
            "required": true,
            "description": "Engine Request Id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "$ref": "#/components/schemas/authResponseError"
          }
        }
      }
    },
    "/api/engines/model/{model}": {
      "get": {
        "summary": "Returns an Engine by its model",
        "tags": [
          "Engine Requests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "model",
            "required": true,
            "description": "Engine model",
            "schema": {
              "type": "string"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/engineResponseBody"
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "id": {
        "properties": {
          "uuid": {
            "type": "string"
          }
        }
      },
      "authResponseError": {
        "description": "Missing Authorization Header"
      },
      "engineResponseBody": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "format": "string"
          },
          "displacement": {
            "type": "int"
          },
          "valve_amount": {
            "type": "int"
          },
          "injectors": {
            "type": "string",
            "format": "string"
          },
          "piston_type": {
            "type": "string",
            "format": "string"
          },
          "camshaft": {
            "type": "string",
            "format": "string"
          },
          "power": {
            "type": "int"
          },
          "forced_induction": {
            "type": "boolean"
          },
          "fuel_type": {
            "type": "string"
          }
        }
      },
      "engineRequestPostBody": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "format": "string"
          },
          "displacement": {
            "type": "int"
          },
          "valve_amount": {
            "type": "int"
          },
          "injectors": {
            "type": "string",
            "format": "string"
          },
          "piston_type": {
            "type": "string",
            "format": "string"
          },
          "camshaft": {
            "type": "string",
            "format": "string"
          },
          "power": {
            "type": "int"
          },
          "forced_induction": {
            "type": "boolean"
          },
          "fuel_type": {
            "type": "string"
          }
        }
      },
      "EngineRequests": {
      }
    }
  }
}